refactor: move preview code to CWHelper
